+++ tests/handlers/test_json.py
@@ -2,15 +2,14 @@
 # MIT License <https://opensource.org/licenses/MIT>
 
 import json
-import ujson
 
 from hypothesis import given
 from hypothesis.strategies import data
 
+import ujson
 import file_config
 
 from .. import config, builder
-
 
 
 @given(config(), data())
@@ -19,3 +18,11 @@ def test_json_dumps(config, data):
     json_content = instance.dumps_json(prefer="json")
     assert isinstance(json_content, str)
     json.loads(json_content)
+
+
+@given(config(), data())
+def test_json_loads(config, data):
+    instance = data.draw(builder.build_config(config))
+    json_content = instance.dumps_json(prefer="json")
+    new_instance = config.loads_json(json_content, prefer="json")
+    assert isinstance(new_instance, config)